requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
requests
python-dotenv
loguru
httpx
pytest
